% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanshiftApplication.R
\name{applyMeanshift}
\alias{applyMeanshift}
\title{Apply Meanshift}
\usage{
applyMeanshift(x, stepsize = 50)
}
\arguments{
\item{x}{-------- data.frame / data.table
format: [Chromosome, Start Position, End Position, X1235 Methylated, X1235 Unmethylated, ...]}

\item{stepsize}{- integer (window size that defines range single sets
can differ from the island center to be in the island)}
}
\value{
a list of lists of indices that make up islands for each island
}
\description{
This function generates a list of lists of islands (as determined by a meanshift
algorithm) for each chromosome from a data.frame or data.table of CpG sites. Expects 
`Chromosome`, Start Position`, and `End Position` columns.
}
