% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanshiftApplication.R
\name{formatIslands}
\alias{formatIslands}
\title{Format Islands}
\usage{
formatIslands(x)
}
\arguments{
\item{x}{- list of dataframe for each chromosome, and list of island indices
columns: [Chromosome1, [20, 21]]}
}
\value{
a list of data.tables for each chromosome with combined islands and 
        unchanged single sets
}
\description{
This function takes in a list with data.frames of the original CpG sites and 
a list of the island indices, separated by chromosome, formatting it into a
single final dataframe with the island indices combined into one row.
Expects original CpG site data.table to have `Chromosome`, `Start Position`,
`End Position`, and methylation read columns -- but no summarized data columns.
}
