% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanshiftApplication.R
\name{getIslands}
\alias{getIslands}
\title{Get Islands}
\usage{
getIslands(file, stepsize = 50)
}
\arguments{
\item{file}{----- string that is pathfile to a table of CpG sites
format: [Chromosome, Start Position, End Position, X1235 Methylated, X1235 Unmethylated, ...]}

\item{stepsize}{- integer of window size for meanshift algorithm
default value = 50}
}
\value{
a data.table which is the original data.table but with the islands (as
        determined by the meanshift algorithm using the given stepsize) added into one row
}
\description{
This function takes in a list with data.frames of the original CpG sites and 
a list of the island indices, separated by chromosome, formatting it into a
single final dataframe with the island indices combined into one row.
Expects original CpG site data.table to have `Chromosome`, `Start Position`,
`End Position`, and methylation read columns -- but no summarized data columns.
}
