% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanshiftApplication.R
\name{getMode}
\alias{getMode}
\title{Get Mode}
\usage{
getMode(x)
}
\arguments{
\item{x}{- list of numeric values}
}
\value{
the statistical mode of the numerical values
}
\description{
This function takes a sequence of numerical values and returns the statistical mode
}
